preprocessing/requirements.txt
matplotlib==3.8.0
pandas==2.1.2
pyserial==3.5

